+++ packages/sdk/src/db/notes.ts
@@ -3,16 +3,28 @@ import {
   AssetTrait,
   IncludedNote,
   NoteTrait,
+  IncludedNoteWithNullifier,
+  IncludedNoteCommitment,
 } from "@nocturne-xyz/primitives";
 import * as JSON from "bigint-json-serialization";
 import { KV, KVStore } from "./kvStore";
+import { numberToStringPadded } from "..";
 
-type NoteAssetKey = string; // Takes form of NOTES_<address>_<id>
-type AllNotes = Map<NoteAssetKey, IncludedNote[]>;
+const NOTES_BY_INDEX_PREFIX = "NOTES_BY_INDEX";
+const NOTES_BY_ASSET_PREFIX = "NOTES_BY_ASSET";
+const NOTES_BY_NULLIFIER_PREFIX = "NOTES_BY_NULLIFIER";
 
-const NOTES_PREFIX = "NOTES";
+// ceil(log10(2^32))
+const MAX_MERKLE_INDEX_DIGITS = 10;
+
+export type AssetKey = string;
+type AllNotes = Map<AssetKey, IncludedNote[]>;
 
 export class NotesDB {
+  // store the following mappings:
+  //  merkleIndex => Note | bigint (note if owned, commitment otherwise)
+  //  asset => merkleIndex[]
+  //  nullifier => merkleIndex
   public kv: KVStore;
 
   constructor(kv: KVStore) {
@@ -26,85 +38,107 @@ export class NotesDB {
    * @param note the note to format
    * @returns key the corresponding key for the note
    */
-  static formatNoteKey(note: IncludedNote): string {
-    return (
-      NotesDB.formatNoteAssetKey(note.asset) + "_" + NoteTrait.sha256(note)
-    );
+  static formatIndexKey(merkleIndex: number): string {
+    return `${NOTES_BY_INDEX_PREFIX}-${numberToStringPadded(
+      merkleIndex,
+      MAX_MERKLE_INDEX_DIGITS
+    )}`;
   }
 
-  /**
-   * Format an `Asset` into a key prefix for the all notes in the KV store for that asset
-   * It produces a key of the form NOTES_<asset.address>_<asset.id>.
-   *
-   * @param asset asset for the note
-   * @returns keyPrefix the key prefix for the asset the note is for
-   */
-  static formatNoteAssetKey(asset: Asset): NoteAssetKey {
-    return (
-      NOTES_PREFIX +
-      "_" +
-      asset.assetType +
-      "_" +
-      asset.assetAddr.toUpperCase() +
-      "_" +
-      asset.id
-    );
+  static parseIndexKey(key: string): number {
+    return parseInt(key.split("-")[1]);
   }
 
-  /**
-   * Parse a note asset key into an `Asset`. Expects a key of form NOTES_<address>_<id>
-   * into the `Asset` the note is for
-   *
-   * @returns asset a
-   * @throws Error if the key is not of the correct form
-   */
-  static parseAssetFromNoteAssetKey(key: NoteAssetKey): Asset {
-    const arr = key.split("_");
-    if (arr.length !== 4 || arr[0] !== NOTES_PREFIX) {
-      throw Error(`Invalid note asset key: "${key}"`);
-    }
+  static formatAssetKey(asset: Asset): string {
+    return `${NOTES_BY_ASSET_PREFIX}-${
+      asset.assetType
+    }-${asset.assetAddr.toUpperCase()}-${asset.id.toString()}`;
+  }
 
+  static parseAssetKey(key: string): Asset {
+    const [_, assetType, assetAddr, id] = key.split("-");
     return {
-      assetType: AssetTrait.parseAssetType(arr[1]),
-      assetAddr: arr[2],
-      id: BigInt(arr[3]),
+      assetType: AssetTrait.parseAssetType(assetType),
+      assetAddr,
+      id: BigInt(id),
     };
   }
 
-  /**
-   * Store a note
-   *
-   * @param note the note to store
-   * @returns true if successful, false otherwise
-   */
-  async storeNote(note: IncludedNote): Promise<boolean> {
-    const key = NotesDB.formatNoteKey(note);
-
-    if (await this.kv.containsKey(key)) {
-      return true;
-    }
-
-    const value = JSON.stringify(note);
-    return await this.kv.putString(key, value);
+  static formatNullifierKey(nullifier: bigint): string {
+    return `${NOTES_BY_NULLIFIER_PREFIX}-${nullifier.toString()}`;
   }
 
-  /**
-   * Get a note
-   * @param key the key of the note to get
-   * @returns note the note, or undefined if not found
-   * @throws Error if the note is found but is not a valid `IncludedNote`
-   */
-  async removeNote(note: IncludedNote): Promise<boolean> {
-    const key = NotesDB.formatNoteKey(note);
-    return await this.kv.remove(key);
-  }
+  async storeNotesAndCommitments(
+    notesAndCommitments: (IncludedNoteWithNullifier | IncludedNoteCommitment)[]
+  ): Promise<void> {
+    // merkleIndex => Note
+    const baseKVs: KV[] = notesAndCommitments.map((noteOrCommitment) => {
+      if (NoteTrait.isNoteNotCommitment(noteOrCommitment)) {
+        const includedNote = noteOrCommitment as IncludedNoteWithNullifier;
+        const note = NoteTrait.toNote(includedNote);
+        return [
+          NotesDB.formatIndexKey(includedNote.merkleIndex),
+          JSON.stringify(note),
+        ];
+      } else {
+        const commitment = noteOrCommitment as IncludedNoteCommitment;
+        return [
+          NotesDB.formatIndexKey(commitment.merkleIndex),
+          commitment.noteCommitment.toString(),
+        ];
+      }
+    });
 
-  async storeNotes(notes: IncludedNote[]): Promise<boolean> {
-    const kvs: KV[] = notes.map((note) => [
-      NotesDB.formatNoteKey(note),
-      JSON.stringify(note),
+    const notes = notesAndCommitments.filter(
+      NoteTrait.isNoteNotCommitment
+    ) as IncludedNoteWithNullifier[];
+
+    // nullifier => merkleIndex
+    const nullifierKVs: KV[] = notes.map((note) => [
+      NotesDB.formatNullifierKey(note.nullifier),
+      NotesDB.formatIndexKey(note.merkleIndex),
     ]);
-    return await this.kv.putMany(kvs);
+
+    // asset => merkleIndex[]
+    const assetKVMap = new Map<AssetKey, string[]>();
+    for (const note of notes) {
+      const assetKey = NotesDB.formatAssetKey(note.asset);
+      let indexKeys = assetKVMap.get(assetKey);
+      if (!indexKeys) {
+        indexKeys = await this.getMerkleIndexKeysForAsset(note.asset);
+      }
+
+      assetKVMap.set(assetKey, [
+        ...indexKeys,
+        NotesDB.formatIndexKey(note.merkleIndex),
+      ]);
+    }
+    const assetKVs: KV[] = Array.from(assetKVMap.entries()).map(
+      ([key, value]) => [key, JSON.stringify(value)]
+    );
+
+    await this.kv.putMany([...baseKVs, ...nullifierKVs, ...assetKVs]);
+  }
+
+  async removeNotesByNullifier(nullifiers: bigint[]): Promise<void> {
+    const nfKeys = nullifiers.map((nullifier) =>
+      NotesDB.formatNullifierKey(nullifier)
+    );
+    const kvs = await this.kv.getMany(nfKeys);
+    const idxKeys = kvs.map(([_nfKey, idxKey]) => idxKey);
+    await this.kv.removeMany([...nfKeys, ...idxKeys]);
+  }
+
+  /**
+   * Get all notes for an asset
+   *
+   * @param asset the asset to get notes for
+   * @returns notes an array of notes for the asset. The array has no guaranteed order.
+   */
+  async getNotesForAsset(asset: Asset): Promise<IncludedNote[]> {
+    const indexKeys = await this.getMerkleIndexKeysForAsset(asset);
+
+    return await this.getNotesByIndexKeys(indexKeys);
   }
 
   /**
@@ -114,38 +148,42 @@ export class NotesDB {
    *          and values are an array of `IncludedNote`s for that asset. The array has no guaranteed order.
    */
   async getAllNotes(): Promise<AllNotes> {
-    const allNotes = new Map<NoteAssetKey, IncludedNote[]>();
+    const allNotes = new Map<AssetKey, IncludedNote[]>();
 
-    const iterPrefix = await this.kv.iterPrefix(NOTES_PREFIX);
-    for await (const [, value] of iterPrefix) {
-      const note = JSON.parse(value);
-      const noteAssetKey = NotesDB.formatNoteAssetKey(note.asset);
+    const iterPrefix = await this.kv.iterPrefix(NOTES_BY_ASSET_PREFIX);
+    for await (const [assetKey, stringifiedIndexKeys] of iterPrefix) {
+      const indexKeys: string[] = JSON.parse(stringifiedIndexKeys);
+      const notes = await this.getNotesByIndexKeys(indexKeys);
 
-      const notesForAsset = allNotes.get(noteAssetKey) ?? [];
-      notesForAsset.push(note);
+      const notesForAsset = allNotes.get(assetKey) ?? [];
+      notesForAsset.push(...notes);
 
-      allNotes.set(noteAssetKey, notesForAsset);
+      if (notesForAsset.length !== 0) {
+        allNotes.set(assetKey, notesForAsset);
+      }
     }
 
     return allNotes;
   }
 
-  /**
-   * Get all notes for an asset
-   *
-   * @param asset the asset to get notes for
-   * @returns notes an array of notes for the asset. The array has no guaranteed order.
-   */
-  async getNotesFor(asset: Asset): Promise<IncludedNote[]> {
-    const noteAssetKey = NotesDB.formatNoteAssetKey(asset);
-
-    const notes = [];
-    const iterPrefix = await this.kv.iterPrefix(noteAssetKey);
-    for await (const [, value] of iterPrefix) {
-      const note = JSON.parse(value);
-      notes.push(note);
+  private async getMerkleIndexKeysForAsset(asset: Asset): Promise<string[]> {
+    const assetKey = NotesDB.formatAssetKey(asset);
+    const value = await this.kv.getString(assetKey);
+    if (!value) {
+      return [];
     }
 
-    return notes;
+    return JSON.parse(value);
+  }
+
+  private async getNotesByIndexKeys(
+    idxKeys: string[]
+  ): Promise<IncludedNote[]> {
+    const kvs = await this.kv.getMany(idxKeys);
+    return kvs.map(([key, value]) => {
+      const merkleIndex = NotesDB.parseIndexKey(key);
+      const note = JSON.parse(value);
+      return NoteTrait.toIncludedNote(note, merkleIndex);
+    });
   }
 }